sony-8974: Disable Metadata storage in buffers.
Legacy sony-msm8974 shall not store metadata in buffers since wrong
code paths would be entered in stagefright that break video processing.

Context:
Storing the metadata of captured frames in the buffer breaks recording
for legacy devices, since "wrong" code paths are followed in this case.
This behaviour is hardcoded by a default parameter in the CameraSource
and CameraSourceTimelapse constructors.

This change allows to falsify TARGET_STORES_METADATA_IN_BUFFERS such
that the metadata storage can be controlled in from device makefiles.

Change-Id: I7d676b20e5ab2474ae966be47112daa06b7d20fd
Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>